feat: create register form
