+++ credit-analyze-portal/src/pages/Register.tsx
@@ -1,0 +1,9 @@
+import { RegisterForm } from '../components/RegisterForm'
+
+export default function Register() {
+  return (
+    <main className="h-screen bg-zinc-50 flex flex-col gap-10 items-center justify-center">
+      <RegisterForm />
+    </main>
+  )
+}
